+++ android/app/src/main/java/com/tracking/MainApplication.java
@@ -1,4 +1,4 @@
-package com.infinity;
+package com.tracking;
 
 import android.app.Application;
 import android.util.Log;
@@ -8,11 +8,12 @@ import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
+import com.facebook.soloader.SoLoader;
 import com.xgfe.reactnativeenv.RCTNativeEnvPackage;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
+import java.util.HashMap;
 
 public class MainApplication extends Application implements ReactApplication {
 
@@ -23,19 +24,25 @@ public class MainApplication extends Application implements ReactApplication {
     }
 
     @Override
-    protected List<ReactPackage> getPackages() {
-        HashMap<String, Object> envs = new HashMap();
-        envs.put("IS_PROD", BuildConfig.IS_PROD);
+        protected List<ReactPackage> getPackages() {
+            HashMap<String, Object> envs = new HashMap();
+            envs.put("IS_PROD", BuildConfig.IS_PROD);
 
-        return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-            new RCTNativeEnvPackage(BuildConfig.class, envs)
-        );
-    }
+            return Arrays.<ReactPackage>asList(
+              new MainReactPackage(),
+                new RCTNativeEnvPackage(BuildConfig.class, envs)
+            );
+        }
   };
 
   @Override
   public ReactNativeHost getReactNativeHost() {
-      return mReactNativeHost;
+    return mReactNativeHost;
+  }
+
+  @Override
+  public void onCreate() {
+    super.onCreate();
+    SoLoader.init(this, /* native exopackage */ false);
   }
 }
