+++ index.android.js
@@ -4,4 +4,4 @@ import {AppRegistry} from 'react-native';
 import setup from './src/setup';
 var React = require('react');
 
-AppRegistry.registerComponent('infinity', setup);
+AppRegistry.registerComponent('tracking', setup);
