+++ ios/tracking/Images.xcassets/AppIcon.appiconset/Contents.json
