+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "infinity",
+  "name": "tracking",
   "version": "0.0.1",
   "private": true,
   "scripts": {
